// src/types/ConnectRole.ts
/**
 * Possible roles in a connect group
 */
export enum ConnectRole {
  ConnectLeader = 'connect-leader',
  RegionalLeader = 'regional-leader',
  ClusterHead = 'cluster-head',
  ConnectMember = 'member',
  ConnectLink = 'connect-link',
  AssistantLeader = 'assistant-leader',
  Admin = 'admin',
  DepartmentHead = 'department-head'
}